Added some functionality to the OngoingCallActivity which will use the bandwidth sdk to make a call.
Created an asynchronous task to perform the call.

Attempted to import the bandwidth sdk and its associated dependencies. Currently does not work. The .jar files included are repackaged to avoid conflicts with the default legacy version that has precedent. The bandwidth sdk is modified to use the repackaged .jar files.